# specs/001-neon-migration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: RESTful API for Todo application with JWT authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://todo-api-production.vercel.app/api/v1
    description: Production server
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Create new user account
      description: Register a new user with email and password
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignup'
            example:
              email: "user@example.com"
              password: "SecurePass123"
              confirm_password: "SecurePass123"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      description: Authenticate user and receive JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: "user@example.com"
              password: "SecurePass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout user
      description: Invalidate refresh token
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user profile
      description: Retrieve authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (expires in 15 minutes)
...
  schemas:
    UserSignup:
      type: object
      required:
        - email
        - password
        - confirm_password
      properties:
        email:
          type: string
          format: email
          description: User's email address (will be lowercase)
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: Password (min 8 chars, at least 1 letter + 1 number)
        confirm_password:
          type: string
          format: password
          description: Password confirmation (must match password)
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          description: User's password
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: JWT refresh token (expires in 7 days)
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token (valid for 15 minutes)
        refresh_token:
          type: string
          description: JWT refresh token (valid for 7 days)
        token_type:
          type: string
          enum: [bearer]
          description: Token type (always "bearer")
        expires_in:
          type: integer
          description: Access token expiry time in seconds
          example: 900
    User:
      type: object
      required:
        - id
        - email
        - role
        - is_verified
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
        email:
          type: string
          format: email
          description: User's email address (lowercase)
        role:
          type: string
          enum: [user, admin]
          description: User's role
        is_verified:
          type: boolean
          description: Whether user's email is verified
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/category
          example: "validation_error"
        message:
          type: string
          description: Human-readable error message
          example: "Email is required"
        details:
          type: object
          description: Additional error details (optional)
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Passwords do not match"
            details: {}
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid or expired token"
            details: {}
...
    Conflict:
      description: Resource conflict (duplicate)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "conflict"
            message: "Email already registered"
            details: {}
    ValidationError:
      description: Validation error (invalid input format)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              email: "Invalid email format"
              password: "Password must be at least 8 characters"